migre todo el codigo a las activities dentro del ui.login, y borre las antiguas :3
