--- Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ActivityBuscar.kt
@@ -1,14 +1,0 @@
-package com.example.filmcast
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-
-class ActivityBuscar : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?){
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_buscar)
-    }
-}
