--- Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/GuardadosActivity.kt
@@ -1,25 +1,0 @@
-package com.example.filmcast
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.filmcast.adapter.PerfilAdapter
-import com.example.filmcast.recycler.PerfilProvider
-
-class GuardadosActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_actores_lista)
-        initRecyclerView()
-    }
-
-    private fun initRecyclerView(){
-        val recyclerView = findViewById<RecyclerView>(R.id.recyclerProfile)
-        recyclerView.layoutManager = LinearLayoutManager(this)
-        recyclerView.adapter = PerfilAdapter(PerfilProvider.perfilList)
-    }
-}
