+++ Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ui/login/MenuActivity.kt
@@ -6,17 +6,22 @@ import android.widget.Button
 import android.widget.ImageView
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.example.filmcast.R
+import com.example.filmcast.adapter.ProyectoAdapter
+import com.example.filmcast.recycler.ProyectoProvider
 
-class menuActivity : AppCompatActivity() {
+class MenuActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContentView(R.layout.activity_menu)
+        initRecyclerView()
 
         val menu_perfil = findViewById<ImageView>(R.id.menu_titulo_perfil)
         menu_perfil.setOnClickListener {
-            val intent = Intent(this, perfilActivity::class.java)
+            val intent = Intent(this, PerfilActivity::class.java)
             startActivity(intent)
         }
         val menu_buscar = findViewById<Button>(R.id.menu_button_buscar)
@@ -30,5 +35,22 @@ class menuActivity : AppCompatActivity() {
             startActivity(intent)
         }
 
+    }
+
+    private fun initRecyclerView() {
+        val recyclerView = findViewById<RecyclerView>(R.id.recyclerProyectos)
+        recyclerView.layoutManager = LinearLayoutManager(
+            this,
+            LinearLayoutManager.HORIZONTAL,
+            false
+        )
+
+        ProyectoProvider.obtenerProyecto { proyectos ->
+            if (proyectos != null) {
+                recyclerView.adapter = ProyectoAdapter(proyectos)
+            } else {
+                println("Error al obtener proyectos")
+            }
+        }
     }
 }
