--- Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ui/login/perfilActivity.kt
@@ -1,17 +1,0 @@
-package com.example.filmcast.ui.login
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import com.example.filmcast.R
-
-class perfilActivity: AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_perfil)
-
-    }
-}
