+++ Aplicaciones de Moviles/app/app/src/main/res/layout/activity_actor.xml
@@ -13,8 +13,7 @@
         android:id="@+id/Perfil_Actor1"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/black"
-        tools:context=".LoginActivity">
+        android:background="@color/black">
 
 
         <androidx.constraintlayout.widget.ConstraintLayout
