+++ Aplicaciones de Moviles/app/app/src/main/res/layout/activity_actores_lista.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.login.MainActivity">
+    tools:context=".ui.login.LoginActivity">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/black"
-        tools:context=".ui.login.MainActivity"
+        tools:context=".ui.login.LoginActivity"
         tools:layout_editor_absoluteX="16dp"
         tools:layout_editor_absoluteY="-194dp">
