+++ Aplicaciones de Moviles/app/app/src/main/res/layout/activity_menu.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.login.menuActivity">
+    tools:context=".ui.login.MenuActivity">
 
     <TextView
         android:id="@+id/menu_titulo"
