+++ Aplicaciones de Moviles/app/app/src/main/res/layout/activity_perfil.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.login.perfilActivity">
+    tools:context=".ui.login.PerfilActivity">
 
 
     <ImageView
