+++ Aplicaciones de Moviles/app/app/src/main/res/layout/activity_register.xml
@@ -62,14 +62,14 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.55"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/et_email">
+        app:layout_constraintTop_toBottomOf="@id/et_password">
 
         <EditText
-            android:id="@+id/et_password"
+            android:id="@+id/et_password1"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:hint="@string/hint_password"
+            android:hint="@string/hint_confirm_password"
             android:inputType="textPassword"
             android:padding="10dp"
             android:textColor="@android:color/white"
@@ -87,12 +87,13 @@
 
 
     <EditText
-        android:id="@+id/et_email"
+        android:id="@+id/et_password"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="30dp"
         android:layout_marginTop="20dp"
-        android:hint="@string/hint_email"
+        android:hint="@string/hint_password"
+        android:inputType="textPassword"
         android:minHeight="48dp"
         android:padding="10dp"
         android:textColor="@android:color/white"
