+++ Aplicaciones de Moviles/app/app/src/main/res/layout/item_perfil.xml
@@ -12,7 +12,7 @@
         android:layout_height="124dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="28dp"
-        android:background="@color/black"
+        android:src="@drawable/anonymous"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
